code/week3/requirements.txt
streamlit==1.29.0
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
plotly==5.18.0
Pillow==10.1.0 